refactoring routes
